+++ docs/example_template.rst
@@ -1,7 +1,7 @@
 .. code:: django
 
   {# Load the tag library #}
-  {% load bootstrap4 %}
+  {% load django_bootstrap4 %}
 
   {# Load CSS and JavaScript #}
   {% bootstrap_css %}
