+++ example/templates/app/base.html
@@ -1,6 +1,6 @@
 {% extends 'app/bootstrap.html' %}
 
-{% load bootstrap4 %}
+{% load django_bootstrap4 %}
 
 {% block bootstrap4_content %}
     <div class="container">
