+++ example/templates/app/home.html
@@ -1,5 +1,5 @@
 {% extends 'app/base.html' %}
-{% load bootstrap4 %}
+{% load django_bootstrap4 %}
 
 {% block title %}django-bootstrap4{% endblock %}
