+++ example/templates/app/misc.html
@@ -1,6 +1,6 @@
 {% extends 'app/base.html' %}
 
-{% load bootstrap4 %}
+{% load django_bootstrap4 %}
 
 {% block title %}
     Miscellaneous
