+++ src/django_bootstrap4/templates/django_bootstrap4/bootstrap4.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-{% load bootstrap4 %}
+{% load django_bootstrap4 %}
 {% if 'use_i18n'|bootstrap_setting %}
     {% load i18n %}
     {% get_current_language as LANGUAGE_CODE %}
