+++ src/django_bootstrap4/templates/django_bootstrap4/pagination.html
@@ -1,4 +1,4 @@
-{% load bootstrap4 %}
+{% load django_bootstrap4 %}
 {% with bpurl=bootstrap_pagination_url|default:"" %}
 
     <ul class="{{ pagination_css_classes }}">
