+++ tests/test_formsets.py
@@ -1,7 +1,7 @@
 from bs4 import BeautifulSoup
 from django.test import TestCase
 
-from bootstrap4.exceptions import BootstrapError
+from django_django_bootstrap4.exceptions import BootstrapError
 
 from .forms import TestForm
 from .utils import render_form, render_formset, render_template_with_form
