+++ tests/test_version.py
@@ -5,8 +5,8 @@ class VersionTest(TestCase):
     """Test presence of package version."""
 
     def test_version(self):
-        import bootstrap4
+        import django_bootstrap4
 
-        version = bootstrap4.__version__
+        version = django_bootstrap4.__version__
         version_parts = version.split(".")
         self.assertTrue(len(version_parts) >= 2)
